commit-msg: Hooks to encourage proper commit message formatting
This hook is to encourage each commit to respect a certain standard:
<type>: <title> (at least 3 words for title)
<new_line>
<description> (at least 15 words)

This would help to have a clear view on how, why and what have been
modified. If the commits are also properly split, that would help to
have atomic change, easily revertable commit and better code/deployment
quality.

A good example of how a log should look like:
https://github.com/torvalds/linux/commits/master