feat: First draft of a script to update all hooks for Cycloid
The 'git init' command do not overwrite the hooks if they have been
updated. Which is a bit frustrating, and can be difficule to deal with
when having many repository using a hook.

This script is a first draft to have this automatised. It has to be
configured to take 2 variables: directory template for git and root
directory of the repository you want to modify.

It then find all the main .git repository, delete the hooks and run an
update.