fix: Updated script with proper binaries, and simply logic
The binaries are now fetch with which, so it doesn't create problems no
matter the shell used by people.

The find and cd commands have been also removed as they were unecessary
and adding complexity to the code. A simple rm and git init with the
right argument proved to be sufficient.

The Cycloid variable name has been modified for a more neutral one.