fix: commit-msg hook wouldn't allow a type with special character
The \w checking the first line wouldn't allow all type of characters.
To make it easier for everyone and not be bothered anymore, this one has
been replaced by \S which allow any non-spacing characters.

This commit comes from
https://github.com/cycloidio/cycloid-hooks/commit/cad62ccaba0c7ce7d98d916a8d2b865443435fa8
but hasen't been cherry-picked because this line changed as well here:
https://github.com/cycloidio/cycloid-hooks/commit/8e7d1a389344cfc5f45d2cd29fffba0ab57c92b7