pre-push: Adding pre-push hook similar to vault pre-commit one
This hook acts as a safety net. The ansible pre-commit should usually be
enough, but it sometimes happen that people working on a branch make
several local/test commits before rebasing and pushing.

This can prove to be dangerous as the various local/test commits can be
done with --no-verify, thus bypassing the ansible pre-commit hook
notifying them that something has been decrypted and commited.

this is why this pre-push hook, makes a final review before a push, to
verify that at no point in time during this work, a file has been
unencrypted and commited.